% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_jlview}
\alias{is_jlview}
\title{Check if an object is a jlview ALTREP vector}
\usage{
is_jlview(x)
}
\arguments{
\item{x}{An R object}
}
\value{
\code{TRUE} if \code{x} is a jlview ALTREP vector, \code{FALSE} otherwise
}
\description{
Check if an object is a jlview ALTREP vector
}
