% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{jlview_release}
\alias{jlview_release}
\title{Explicitly release a jlview object}
\usage{
jlview_release(x)
}
\arguments{
\item{x}{A jlview ALTREP vector}
}
\value{
Invisible \code{NULL}
}
\description{
Unpins the Julia array immediately, freeing memory without waiting for
R's garbage collector. After release, accessing the data will error.
}
